app/xcf: fix #12980 protection against invalid xcf offset
In `xcf_load_image` when reading the layer offset in an xcf file, we
did not check whether `xcf_read_offset read` the correct amount of
bytes.
We also did not initialize the offset parameter to 0 in
`xcf_read_offset`, which meant that if 0 bytes were read, the offset
could have a random value.

To fix this we initialize *data to 0 in `xcf_read_offset` and check
the amount of bytes returned by the call to  `xcf_read_offset`.
If it is less than the size of the offset variable we stop reading
layers.