core: Use selection when updating GimpBrushClipboard
Resolves #10046
After 1b64fdf5, gimp_get_clipboard_object () no longer returns a merged version of the selected area.
This affected the Clipboard Brush and Clipboard Mask,
causing them to always use the entire image instead.

To restore the 2.10 behavior, we check if the pasted
image has a selection. If so, then we call 
gimp_selection_extract () to get the buffer of the
selected area. We also need to unreference this
buffer afterwards, so a boolean check is added once
the buffer has been copied over to the brush.