plug-ins/help: fix #13049 Help progress not closing
We had to revert our previous fix for #13049 due to other Help related
issues. When there was an error condition when searching or dowloading
the xml help index, the progress widget was not closed and stayed
in the status bar until the Cancel button was clicked.

Fix this by properyly calling `_gimp_help_progress_finish` when tere
is an error condition.