libgimpbase: lower GIMP_MIN_RESOLUTION.
As per #13078, this is required to allow adding the Mil unit commonly
used in electronics. This minimum resolution is not only used for pixel
density, it's actually also used for minimum dimensions (which was our
problem here).

Now the minimum image density is 0.001 pixels per inch, and minimum unit
dimension is 0.001 inches (1 mil).