config, gui: Restore "Show Tooltips" option in Preferences
In b3620517, we removed the option to hide or show
tooltips when hovering over an icon.
The rationale given in #1965 appears to be related to
the added complexity of the option. A suggestion
was given to hardcode the tooltips with an opacity of
0 in the CSS to resolve the issue.

This patch restores the option to toggle tooltips in the
Preferences Dialogue, but simply writes the suggested
CSS as part of themes.css rather than the original,
more complex implementation. This should hopefully
satisfy users who feel this change was a regression
from 2.10, while keeping the implementation simple.